% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanLines.R
\name{CleanLines}
\alias{CleanLines}
\title{Clean lines (SpatialLinesDataFrame) for Steiner tree}
\usage{
CleanLines(x)
}
\arguments{
\item{x}{SpatialLinesDataFrame (network)}
}
\value{
SLDF ready to use in v.net and v.net.steiner
}
\description{
This function will remove unchained parts of the lines layer
that can cause issues when executing v.net.steiner. After removing those parts,
the layer is imported into grass to execute v.clean and v.build to repare
topology issues#'
}
\examples{
\dontrun{
setGRASS(gisBase = "/usr/lib/grass78", epsg= 25829)
data("l")
lc <- CleanLines(l[1:100,])
mapview(l[1:100,])+lc
}
}
