% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\docType{data}
\name{l}
\alias{l}
\title{l --> network (sldf)}
\description{
l --> network (sldf)
}
\references{
\url{https://www.openstreetmap.org/#map=9/42.7531/-7.6602}
}
\author{
OSM{https://www.openstreetmap.org/#map=9/42.7531/-7.6602}
}
\keyword{data}
