% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\docType{data}
\name{p}
\alias{p}
\title{p --> points (spdf)}
\description{
p --> points (spdf)
}
\references{
\url{https://www.openstreetmap.org/#map=9/42.7531/-7.6602}
}
\author{
OSM{https://www.openstreetmap.org/#map=9/42.7531/-7.6602}
}
\keyword{data}
