% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnifyLines.R
\name{UnifyLines}
\alias{UnifyLines}
\title{UnifyLines (after steiner)}
\usage{
UnifyLines(l)
}
\arguments{
\item{l}{sldf to be unified}
}
\value{
sldf unified into one feature
}
\description{
The algorith transform an SLDF with multiple and overlaping lines
into a unique line that works as a clean network to caltulate the final
Total Steiner Tree
}
\examples{
\dontrun{
data("l")
ls <- list()
for(i in 1:length(l[1:100,])){ls[i] <- l[i,]}
l1 <- spatials2one(ls)

# multiply x2 the lines to tryout UnifyLines
l2 <- rbind(l1,l1)

# UnifyLines (network)
UnifyLines(l2)
}
}
