% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setGRASS.R
\name{setGRASS}
\alias{setGRASS}
\title{Function to initialize rgrass}
\usage{
setGRASS(
  gisBase = "/usr/lib/grass78",
  home = tempdir(),
  gisDbase = tempdir(),
  location = "grassdata",
  mapset = "PERMANENT",
  override = TRUE,
  flags = "c",
  epsg = 25829
)
}
\arguments{
\item{gisBase}{Location of the grass78 library}

\item{home}{Temporal file by default}

\item{gisDbase}{Temporal file by default}

\item{location}{Main folder for importing and processing}

\item{mapset}{Subfolder for importing and processgin}

\item{flags}{Options for rgrass7 package}

\item{epsg}{By default set to 25829}

\item{overrride}{TRUE by default}
}
\value{
There is no output, just setting the grass paths, flags...
}
\description{
Function to initialize rgrass
}
\examples{
setGRASS()
\dontrun{
setGRASS(gisBase = "/usr/lib/grass78", epsg= 25829)
}
}
