% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteinerTree.R
\name{SteinerTree}
\alias{SteinerTree}
\title{SteinerTree function using v.net.steiner from GRASS as core function
It contains the v.net function to connect points to lines using threshold}
\usage{
SteinerTree(l, p, th)
}
\arguments{
\item{l}{Network to calculate Steiner Tree (SLDF)}

\item{p}{Points to be conected by Steiner Tree (SPDF)}

\item{th}{Threshold --> max distance of points to be connecte to lines}
}
\value{

}
\description{
SteinerTree function using v.net.steiner from GRASS as core function
It contains the v.net function to connect points to lines using threshold
}
